call queries up at the root
